// services/auditd/tests/generated/v1/auditd.ts
export interface AuditEvent {
  sourceEventId: string;
  occurredAt: Date | undefined;
  attribution: AuditAttribution | undefined;
  partition: AuditPartition | undefined;
  traceId: string;
  spanId: string;
  tenantMutation?: TenantMutationAuditDetail | undefined;
  workspaceMutation?: WorkspaceMutationAuditDetail | undefined;
  identitySession?: IdentitySessionAuditDetail | undefined;
  packageDeclaration?: PackageDeclarationAuditDetail | undefined;
  appMutation?: AppMutationAuditDetail | undefined;
  configurationPublication?: ConfigurationPublicationAuditDetail | undefined;
  secretPublication?: SecretPublicationAuditDetail | undefined;
  projectionMutation?: ProjectionMutationAuditDetail | undefined;
  placementMutation?: PlacementMutationAuditDetail | undefined;
  workloadMutation?: WorkloadMutationAuditDetail | undefined;
  runMutation?: RunMutationAuditDetail | undefined;
}

export interface AuditAttribution {
  operatorCommonName?: string | undefined;
  workloadSubject?: string | undefined;
  invocation?: InvocationAuditAttribution | undefined;
}

export interface InvocationAuditAttribution {
  actorPrincipalId: string;
  attachedAccountPrincipalId: string;
  workloadSubject: string;
}

export interface AuditPartition {
  global?: GlobalAuditPartition | undefined;
  tenant?: TenantAuditPartition | undefined;
}

export interface GlobalAuditPartition {
}
// ...
export interface TenantMutationAuditDetail {
  action: TenantMutationAction;
  resourceRevision: bigint;
  resultingState: TenancyResourceState;
}

export interface WorkspaceMutationAuditDetail {
  workspaceId: string;
  action: WorkspaceMutationAction;
  resourceRevision: bigint;
  resultingState: TenancyResourceState;
}
// ...
export const AuditEvent: MessageFns<AuditEvent> = {
// ...
  fromJSON(object: any): AuditEvent {
    return {
      sourceEventId: isSet(object.sourceEventId) ? globalThis.String(object.sourceEventId) : "",
      occurredAt: isSet(object.occurredAt) ? fromJsonTimestamp(object.occurredAt) : undefined,
      attribution: isSet(object.attribution) ? AuditAttribution.fromJSON(object.attribution) : undefined,
      partition: isSet(object.partition) ? AuditPartition.fromJSON(object.partition) : undefined,
      traceId: isSet(object.traceId) ? globalThis.String(object.traceId) : "",
      spanId: isSet(object.spanId) ? globalThis.String(object.spanId) : "",
      tenantMutation: isSet(object.tenantMutation)
        ? TenantMutationAuditDetail.fromJSON(object.tenantMutation)
        : undefined,
      workspaceMutation: isSet(object.workspaceMutation)
        ? WorkspaceMutationAuditDetail.fromJSON(object.workspaceMutation)
        : undefined,
      identitySession: isSet(object.identitySession)
        ? IdentitySessionAuditDetail.fromJSON(object.identitySession)
        : undefined,
      packageDeclaration: isSet(object.packageDeclaration)
        ? PackageDeclarationAuditDetail.fromJSON(object.packageDeclaration)
        : undefined,
      appMutation: isSet(object.appMutation) ? AppMutationAuditDetail.fromJSON(object.appMutation) : undefined,
      configurationPublication: isSet(object.configurationPublication)
        ? ConfigurationPublicationAuditDetail.fromJSON(object.configurationPublication)
        : undefined,
      secretPublication: isSet(object.secretPublication)
        ? SecretPublicationAuditDetail.fromJSON(object.secretPublication)
        : undefined,
      projectionMutation: isSet(object.projectionMutation)
        ? ProjectionMutationAuditDetail.fromJSON(object.projectionMutation)
        : undefined,
      placementMutation: isSet(object.placementMutation)
        ? PlacementMutationAuditDetail.fromJSON(object.placementMutation)
        : undefined,
      workloadMutation: isSet(object.workloadMutation)
        ? WorkloadMutationAuditDetail.fromJSON(object.workloadMutation)
        : undefined,
      runMutation: isSet(object.runMutation) ? RunMutationAuditDetail.fromJSON(object.runMutation) : undefined,
    };
  },
// ...
};

function createBaseAuditAttribution(): AuditAttribution {
  return { operatorCommonName: undefined, workloadSubject: undefined, invocation: undefined };
}

export const AuditAttribution: MessageFns<AuditAttribution> = {
  encode(message: AuditAttribution, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.operatorCommonName !== undefined) {
      writer.uint32(10).string(message.operatorCommonName);
    }
    if (message.workloadSubject !== undefined) {
      writer.uint32(18).string(message.workloadSubject);
    }
    if (message.invocation !== undefined) {
      InvocationAuditAttribution.encode(message.invocation, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AuditAttribution {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAuditAttribution();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.operatorCommonName = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.workloadSubject = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.invocation = InvocationAuditAttribution.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AuditAttribution {
    return {
      operatorCommonName: isSet(object.operatorCommonName) ? globalThis.String(object.operatorCommonName) : undefined,
      workloadSubject: isSet(object.workloadSubject) ? globalThis.String(object.workloadSubject) : undefined,
      invocation: isSet(object.invocation) ? InvocationAuditAttribution.fromJSON(object.invocation) : undefined,
    };
  },

  toJSON(message: AuditAttribution): unknown {
    const obj: any = {};
    if (message.operatorCommonName !== undefined) {
      obj.operatorCommonName = message.operatorCommonName;
    }
    if (message.workloadSubject !== undefined) {
      obj.workloadSubject = message.workloadSubject;
    }
    if (message.invocation !== undefined) {
      obj.invocation = InvocationAuditAttribution.toJSON(message.invocation);
    }
    return obj;
  },

  create(base?: DeepPartial<AuditAttribution>): AuditAttribution {
    return AuditAttribution.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<AuditAttribution>): AuditAttribution {
    const message = createBaseAuditAttribution();
    message.operatorCommonName = object.operatorCommonName ?? undefined;
    message.workloadSubject = object.workloadSubject ?? undefined;
    message.invocation = (object.invocation !== undefined && object.invocation !== null)
      ? InvocationAuditAttribution.fromPartial(object.invocation)
      : undefined;
    return message;
  },
};
// ...
export const AuditPartition: MessageFns<AuditPartition> = {
  encode(message: AuditPartition, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.global !== undefined) {
      GlobalAuditPartition.encode(message.global, writer.uint32(10).fork()).join();
    }
    if (message.tenant !== undefined) {
      TenantAuditPartition.encode(message.tenant, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AuditPartition {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAuditPartition();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.global = GlobalAuditPartition.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.tenant = TenantAuditPartition.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AuditPartition {
    return {
      global: isSet(object.global) ? GlobalAuditPartition.fromJSON(object.global) : undefined,
      tenant: isSet(object.tenant) ? TenantAuditPartition.fromJSON(object.tenant) : undefined,
    };
  },

  toJSON(message: AuditPartition): unknown {
    const obj: any = {};
    if (message.global !== undefined) {
      obj.global = GlobalAuditPartition.toJSON(message.global);
    }
    if (message.tenant !== undefined) {
      obj.tenant = TenantAuditPartition.toJSON(message.tenant);
    }
    return obj;
  },

  create(base?: DeepPartial<AuditPartition>): AuditPartition {
    return AuditPartition.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<AuditPartition>): AuditPartition {
    const message = createBaseAuditPartition();
    message.global = (object.global !== undefined && object.global !== null)
      ? GlobalAuditPartition.fromPartial(object.global)
      : undefined;
    message.tenant = (object.tenant !== undefined && object.tenant !== null)
      ? TenantAuditPartition.fromPartial(object.tenant)
      : undefined;
    return message;
  },
};
// ...
export const TenantMutationAuditDetail: MessageFns<TenantMutationAuditDetail> = {
  encode(message: TenantMutationAuditDetail, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.action !== 0) {
      writer.uint32(8).int32(message.action);
    }
    if (message.resourceRevision !== 0n) {
      if (BigInt.asUintN(64, message.resourceRevision) !== message.resourceRevision) {
        throw new globalThis.Error("value provided for field message.resourceRevision of type uint64 too large");
      }
      writer.uint32(16).uint64(message.resourceRevision);
    }
    if (message.resultingState !== 0) {
      writer.uint32(24).int32(message.resultingState);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TenantMutationAuditDetail {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTenantMutationAuditDetail();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.action = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.resourceRevision = reader.uint64() as bigint;
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.resultingState = reader.int32() as any;
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TenantMutationAuditDetail {
    return {
      action: isSet(object.action) ? tenantMutationActionFromJSON(object.action) : 0,
      resourceRevision: isSet(object.resourceRevision) ? BigInt(object.resourceRevision) : 0n,
      resultingState: isSet(object.resultingState) ? tenancyResourceStateFromJSON(object.resultingState) : 0,
    };
  },

  toJSON(message: TenantMutationAuditDetail): unknown {
    const obj: any = {};
    if (message.action !== 0) {
      obj.action = tenantMutationActionToJSON(message.action);
    }
    if (message.resourceRevision !== 0n) {
      obj.resourceRevision = message.resourceRevision.toString();
    }
    if (message.resultingState !== 0) {
      obj.resultingState = tenancyResourceStateToJSON(message.resultingState);
    }
    return obj;
  },

  create(base?: DeepPartial<TenantMutationAuditDetail>): TenantMutationAuditDetail {
    return TenantMutationAuditDetail.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<TenantMutationAuditDetail>): TenantMutationAuditDetail {
    const message = createBaseTenantMutationAuditDetail();
    message.action = object.action ?? 0;
    message.resourceRevision = object.resourceRevision ?? 0n;
    message.resultingState = object.resultingState ?? 0;
    return message;
  },
};